Increase the number of repositories returned in projects
With organizations that have over 30 repos it can be impossible to create tickets as not all repositories are returned.

Increasing the amount of results returned is an interim solution, which may get unwieldy, but that can be solved by adding a filter box or something into the UI later. For now this should help a large number of cases.